+++ src/Dynatrace.Net/Configuration/MobileSymbolication/Models/SymbolFile.cs
@@ -5,7 +5,7 @@
 		public string ApplicationName { get; set; }
 		public int? Size { get; set; }
 		public AppIdentifier AppId { get; set; }
-		public int? UploadTimestamp { get; set; }
+		public long? UploadTimestamp { get; set; }
 		public bool? Pinned { get; set; }
 	}
 }
