+++ src/Dynatrace.Net/Environment/Events/Models/EventQueryResult.cs
@@ -4,12 +4,12 @@ namespace Dynatrace.Net.Environment.Events.Models
 {
 	public class EventQueryResult
 	{
-		public int? NextEventStartTms { get; set; }
-		public int? NextEventId { get; set; }
+		public long? NextEventStartTms { get; set; }
+		public long? NextEventId { get; set; }
 		public string NextCursor { get; set; }
-		public int? From { get; set; }
-		public int? To { get; set; }
-		public int? TotalEventCount { get; set; }
+		public long? From { get; set; }
+		public long? To { get; set; }
+		public long? TotalEventCount { get; set; }
 		public IEnumerable<EventRestEntry> Events { get; set; }
 	}
 }
