+++ src/Dynatrace.Net/Environment/ProblemsV1/Models/ProblemCloseResult.cs
@@ -4,7 +4,7 @@
 	{
 		public string ProblemId { get; set; }
 		public ProblemComment Comment { get; set; }
-		public int? CloseTimestamp { get; set; }
+		public long? CloseTimestamp { get; set; }
 		public bool? Closing { get; set; }
 	}
 }
