+++ src/Dynatrace.Net/Environment/ProblemsV1/Models/ProblemComment.cs
@@ -3,7 +3,7 @@
 	public class ProblemComment
 	{
 		public string Id { get; set; }
-		public int? CreatedAtTimestamp { get; set; }
+		public long? CreatedAtTimestamp { get; set; }
 		public string Content { get; set; }
 		public string UserName { get; set; }
 		public string Context { get; set; }
