+++ src/Dynatrace.Net/Environment/ProblemsV2/Models/CommentDto.cs
@@ -3,7 +3,7 @@
 	public class CommentDto
 	{
 		public string AuthorName { get; set; }
-		public int? CreatedAtTimestamp { get; set; }
+		public long? CreatedAtTimestamp { get; set; }
 		public string Context { get; set; }
 		public string Id { get; set; }
 		public string Content { get; set; }
