+++ src/Dynatrace.Net/Environment/ProblemsV2/Models/ProblemCloseResult.cs
@@ -3,7 +3,7 @@
 	public class ProblemCloseResult
 	{
 		public bool? Closing { get; set; }
-		public int? CloseTimestamp { get; set; }
+		public long? CloseTimestamp { get; set; }
 		public string ProblemId { get; set; }
 		public CommentDto Comment { get; set; }
 	}
