+++ src/Dynatrace.Net/Environment/SyntheticV1/Models/SyntheticMonitorStepResult.cs
@@ -3,7 +3,7 @@
 	public class SyntheticMonitorStepResult
 	{
 		public int? Id { get; set; }
-		public int? StartTimestamp { get; set; }
+		public long? StartTimestamp { get; set; }
 		public int? ResponseTimeMillis { get; set; }
 		public SyntheticMonitorError Error { get; set; }
 	}
