+++ src/Dynatrace.Net/Environment/SyntheticV1/Models/ThirdPartyEventResolvedNotification.cs
@@ -4,6 +4,6 @@
 	{
 		public string TestId { get; set; }
 		public string EventId { get; set; }
-		public int? EndTimestamp { get; set; }
+		public long? EndTimestamp { get; set; }
 	}
 }
