+++ src/Dynatrace.Net/Environment/UserSessions/Models/UserSessionSyntheticEvent.cs
@@ -5,7 +5,7 @@
 		public string Name { get; set; }
 		public string SyntheticEventId { get; set; }
 		public int? SequenceNumber { get; set; }
-		public int? Timestamp { get; set; }
+		public long? Timestamp { get; set; }
 		public string Type { get; set; }
 		public int? ErrorCode { get; set; }
 		public string ErrorName { get; set; }
